Update the linked test case hmm.pdf to point to the intended version
The test case was changed in https://github.com/mozilla/pdf.js/commit/1faca190216deed1ef6084697d04a1d90621e210 because the original file was not available anymore. However, its hash was also changed, meaning that we do not test the intended version anymore.

This patch makes sure that we test the intented version by reverting to
the original hash and using a link, also pointing to the Internet
Archive, with the original file.